Silent a gcc-4.5 warning about a 'case' label not in enumerated type
The enum in question is a set of flags that are or'ed together, so obviously
the or'ed combination should not be defined in the enum (otherwise there would
be no point in or'ing anything).
Anyway ... casted the enum to unsigned to silent -Wswitch with gcc-4.5

See http://gcc.gnu.org/bugzilla/show_bug.cgi?id=46457 for a similar complaint.

+++ src/port.c
@@ -121,7 +121,7 @@ port_destroy(port_s p)
 const char *
 port_strkind(portkind k)
 {
-  switch(k) {
+  switch((unsigned int)k) {
     case PORT_IN|PORT_DATA|PORT_STATIC:		return "data in";
     case PORT_IN|PORT_EVENT|PORT_STATIC:	return "event in";
     case PORT_OUT|PORT_DATA|PORT_STATIC:	return "data out";
